changed name of SetLocation to LocationSetter
